Setup the initial layout for corne lp aka crkbd keyboard
This is the very first change to start to support corne lp, aka crkbd keyboard.

the keymap is adopting my current latest layout for split keyboard.

it has the minimal to get a working keyboard to use.

will need to improve once problems occured.

e.g. one known issue is the z key is using LSFT_T(KC_Z) which is bad, because then type fast, e.g. hyangzhang will end up as hyangHang, because z and h is so fast, hence z is recognized as shift instead of z and h.
hence need to follow the same approach a is doing, aka a customized tap dance to overcome this problem.